+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "the-boct",
-  "version": "4.0.0-alpha-7.1",
+  "version": "4.0.0-alpha-7.2",
   "license": "MIT",
   "homepage": "https://the-boct.github.io/",
   "private": true,
@@ -36,5 +36,8 @@
   },
   "prettier": {
     "printWidth": 240
+  },
+  "babel": {
+
   }
 }
