+++ public/index.html
@@ -9,11 +9,10 @@
     <meta name="description" content="Meet the BOcT" />
     <link rel="icon" href="assets/robot.ico" />
 
-    <link rel="preconnect" href="https://fonts.googleapis.com/" crossorigin />
+    <link rel="preconnect" href="https://www.googletagmanager.com/" />
     <link rel="preconnect" href="https://fonts.gstatic.com/" crossorigin />
     <link rel="preconnect" href="https://www.google-analytics.com/" crossorigin />
 
-    <!--PWA-->
     <meta name="application-name" content="The-BOcT" />
     <meta name="theme-color" content="#E64C63" />
     <link rel="apple-touch-icon" sizes="128x128" href="assets/robot128.png" />
@@ -21,8 +20,6 @@
   </head>
   <body>
     <div id="Main"></div>
-
-    <noscript>You need to enable JavaScript to run this app.</noscript>
 
     <script defer src="https://www.googletagmanager.com/gtag/js?id=UA-172329807-1"></script>
     <script defer>
