+++ src/lib/chat/units.ts
@@ -1,73 +1,3 @@
-import { replyRandom } from "./small-functions";
-//AH BUGS!!
-
-function unit_convert(usertyped: string): string {
-  const ucs_data: string[] = usertyped.split(" ");
-
-  if (ucs_data.length === 4) {
-    let uVal = parseFloat(ucs_data[1]);
-    let uValLeng = String(uVal).replace(".", " ").length;
-    let uUnit = ucs_data[1].slice(uValLeng);
-    return TheConverter(uVal, uUnit, ucs_data[3]);
-  }
-  else {
-    return "Type in this Format: conv 37cm to m";
-  }
-}
-
-function TheConverter(x: number, a: string, b: string): string {
-  let from_unit = getUnit(a);
-  let to_unit = getUnit(b);
-
-  if (from_unit.Available && to_unit.Available) {
-    if (from_unit.Category == to_unit.Category) {
-      if (from_unit.con_factor && to_unit.con_factor) {
-        let result = x * (from_unit.con_factor / to_unit.con_factor);
-        return `${result} ${b}`;
-      }
-      else if (from_unit.con_trnsTo && to_unit.con_trnsFro) {
-        let SIval = from_unit.con_trnsTo(x);
-        let result = to_unit.con_trnsFro(SIval);
-        return `${result}${b}`;
-      }
-    }
-    else {
-      return replyRandom(["-_-", "Conversions do not work that way", "Maybe, My Scripts are incomplete. Couldn't Convert"]);
-    }
-  }
-
-  return replyRandom(["That doesn't make sense", "Something is very Wrong here."]);
-}
-
-function getUnit(u: string) {
-  const Categories = [Lengths, Areas, Mass, Temperatures, Volume];
-  let unit_details: {
-    Available: boolean;
-    Category?: number;
-    con_factor?: number;
-    con_trnsTo?: any;
-    con_trnsFro?: any;
-  } = {
-    Available: false,
-  };
-  let i = 0;
-
-  while (i < Categories.length && !unit_details.Available) {
-    Object.keys(Categories[i]).forEach((j) => {
-      if (Categories[i][j].unit.includes(u)) {
-        let pre_unit_detail = {
-          Available: true,
-          Category: i,
-        };
-        unit_details = Object.assign(pre_unit_detail, Categories[i][j]);
-      }
-    });
-    i++;
-  }
-
-  return unit_details;
-}
-
 // 1 x is equal to con_factor y (x is a unit; y is respective SI/base unit)
 //Ex: 1 mm is equal to 1/1000 m
 const Lengths = {
@@ -212,9 +142,11 @@ const Volume = {
   },
 };
 
+const Categories =  [Lengths, Areas, Temperatures, Volume, Mass];
+export default Categories;
+
 /*
 const Digital = {}
-
 const Prefixs = {
   uni: {
     unit: [''],
@@ -244,7 +176,4 @@ const Prefixs = {
     unit: ['T','tera','Tera'],
     con_factor: 1e+12
   }
-}
-*/
-
-export default unit_convert;
+}*/
